requirements.txt
pyrogram
supabase
python-dotenv